Add language to GameStarter, Move client-only setup
