+++ theLastHourEscapeRoom/src/starter/LastHourClient.java
@@ -1,5 +1,9 @@
 package starter;
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input;
+import com.badlogic.gdx.audio.Music;
+import com.badlogic.gdx.graphics.Color;
 import contrib.components.CollideComponent;
 import contrib.entities.CharacterClass;
 import contrib.entities.HeroBuilder;
@@ -18,14 +22,28 @@ import core.components.PositionComponent;
 import core.configuration.KeyboardConfig;
 import core.game.ClientStarter;
 import core.game.PreRunConfiguration;
+import core.language.Language;
+import core.language.Localization;
 import core.network.ConnectionListener;
 import core.network.messages.s2c.EntitySpawnEvent;
 import core.utils.CursorUtil;
 import core.utils.Tuple;
 import core.utils.components.draw.DrawComponentFactory;
+import core.utils.components.draw.TextureGenerator;
+import core.utils.components.draw.TextureMap;
+import core.utils.components.draw.shader.ColorGradeShader;
+import core.utils.components.draw.shader.HueRemapShader;
+import core.utils.components.draw.shader.ShaderList;
 import core.utils.components.path.SimpleIPath;
+
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+
+import core.utils.settings.ButtonBindingSetting;
+import core.utils.settings.ClientSettings;
+import core.utils.settings.DescriptionSetting;
+import core.utils.settings.SectionDividerSetting;
 import level.LastHourLevel;
 import level.LastHourLevelClient;
 import modules.computer.ComputerFactory;
@@ -40,12 +58,16 @@ import util.ui.BlackFadeCutscene;
 /** The main class for the Multiplayer Client for development and testing purposes. */
 public final class LastHourClient {
 
+  private static final String BACKGROUND_MUSIC = "sounds/forest_bgm.wav";
+  private static Music backgroundMusic;
+
   /**
    * Main method to start the standalone dev client (bypasses the main menu).
    *
    * @param args command line arguments
    */
   public static void main(String[] args) {
+    Game.localization().currentLanguage(Language.EN);
     starter().apply();
     Game.run();
   }
@@ -87,7 +109,10 @@ public final class LastHourClient {
     Game.remove(AttributeBarSystem.class);
     Game.add(new ComputerStateSyncSystem());
 
-    TheLastHour.setupClient();
+    initLocalization();
+    setupMusic();
+    staticRenderTextures();
+    registerSettings();
 
     Game.network()
         .addConnectionListener(
@@ -245,5 +270,122 @@ public final class LastHourClient {
       // Best-effort: if anything goes wrong, fall through to the default texture loading
       // and let it fail loudly (the resulting visible error is clearer than a partial spawn).
     }
+  }
+
+  private static void initLocalization() {
+    Localization localization = Game.localization();
+    localization.registerTranslationFile(Language.DE, "language/de.json");
+    localization.registerTranslationFile(Language.EN, "language/en.json");
+  }
+
+  private static final String T_SETTINGS_CONTROLS_HEADER = "settings.controls_header";
+  private static final String T_SETTINGS_CONTROLS_DESCRIPTION = "settings.controls_description";
+  private static final String T_SETTINGS_PAUSE = "settings.pause";
+  private static final String T_SETTINGS_INTERACT = "settings.interact";
+  private static final String T_SETTINGS_INVENTORY = "settings.inventory";
+  private static final String T_SETTINGS_INVENTORY_DESCRIPTION = "settings.inventory_description";
+
+  /** Registers additional client settings. */
+  private static void registerSettings() {
+    ClientSettings.registerSetting(new SectionDividerSetting(T_SETTINGS_CONTROLS_HEADER));
+    ClientSettings.registerSetting(
+      new DescriptionSetting(T_SETTINGS_CONTROLS_DESCRIPTION, Input.Keys.E));
+    ClientSettings.registerSetting(new ButtonBindingSetting(T_SETTINGS_PAUSE, Input.Keys.P, false));
+    ClientSettings.registerSetting(
+      new ButtonBindingSetting(T_SETTINGS_INTERACT, Input.Keys.E, false));
+    ClientSettings.registerSetting(
+      new ButtonBindingSetting(T_SETTINGS_INVENTORY, Input.Keys.I, false));
+    ClientSettings.registerSetting(
+      new DescriptionSetting(T_SETTINGS_INVENTORY_DESCRIPTION, Input.Buttons.RIGHT));
+  }
+
+  private static final List<Tuple<String, Color>> USB_TEXTURES =
+    List.of(
+      Tuple.of("items/usb-side-green.png", Color.GREEN),
+      Tuple.of("items/usb-side-blue.png", Color.BLUE),
+      Tuple.of("items/usb-side-yellow.png", Color.YELLOW));
+
+  /** Statically renders the needed textures. */
+  private static void staticRenderTextures() {
+    String basePath = "items/usb-side-red.png";
+    float baseHue = 0.0f;
+
+    for (Tuple<String, Color> usbTexture : USB_TEXTURES) {
+      ShaderList shaderList = new ShaderList();
+
+      String outTexturePath = usbTexture.a();
+      Color color = usbTexture.b();
+      float[] hsv = new float[3];
+      shaderList.add("hueRemap", new HueRemapShader(baseHue, color.toHsv(hsv)[0] / 360f));
+
+      TextureGenerator.registerRenderShaderTexture(basePath, outTexturePath, shaderList);
+    }
+
+    // Invert the keyboard / mouse input-prompt spritesheet so the white-on-transparent icons
+    // read clearly against the dark HUD text used in this game.
+    String keyboardPromptPath = "hud/input/keyboard_mouse.png";
+    ShaderList invertShaders = new ShaderList();
+    invertShaders.add("invert", new ColorGradeShader().invert(true));
+    TextureGenerator.registerRenderShaderTexture(
+      keyboardPromptPath, keyboardPromptPath, invertShaders);
+
+    // Put the first frame of the idle animation of the rogue and the char03 characters into a
+    // special texture in "@gen/char03.png" and "@gen/rogue.png", so they can be used for Dialogs
+    // without needing to parse the spritesheet again.
+    registerCharacterPortrait(CharacterClass.THE_LAST_HOUR_ROGUE, ROGUE_PORTRAIT_PATH);
+    registerCharacterPortrait(CharacterClass.THE_LAST_HOUR_CHAR03, CHAR03_PORTRAIT_PATH);
+  }
+
+  /** Path of the generated portrait texture for the Rogue character. */
+  public static final String ROGUE_PORTRAIT_PATH = "@gen/rogue.png";
+
+  /** Path of the generated portrait texture for the Char03 character. */
+  public static final String CHAR03_PORTRAIT_PATH = "@gen/char03.png";
+
+  /** Width / height in pixels of a single frame in the character spritesheets. */
+  private static final int CHARACTER_FRAME_SIZE = 32;
+
+  private static final int CHARACTER_FRAME_PADDING = 8;
+
+  /**
+   * Extracts the first frame from the given character's spritesheet and registers it as a
+   * standalone texture in the {@link TextureMap} under {@code outPath}.
+   *
+   * @param characterClass character whose sprite sheet should be sampled
+   * @param outPath virtual texture-map output path for the generated portrait
+   */
+  private static void registerCharacterPortrait(CharacterClass characterClass, String outPath) {
+    String sheetPath = characterClass.textures().pathString();
+    if (!sheetPath.endsWith(".png")) {
+      sheetPath = sheetPath + "/" + sheetPath.substring(sheetPath.lastIndexOf('/') + 1) + ".png";
+    }
+    TextureGenerator.registerSpritesheetRegionTexture(
+      sheetPath,
+      CHARACTER_FRAME_PADDING,
+      CHARACTER_FRAME_PADDING,
+      CHARACTER_FRAME_SIZE - CHARACTER_FRAME_PADDING * 2,
+      CHARACTER_FRAME_SIZE - CHARACTER_FRAME_PADDING * 2,
+      outPath);
+  }
+
+  /**
+   * Initializes and starts the background music for the game, and sets up listeners to adjust the
+   * volume based on client settings changes.
+   */
+  private static void setupMusic() {
+    backgroundMusic = Gdx.audio.newMusic(Gdx.files.internal(BACKGROUND_MUSIC));
+    backgroundMusic.setLooping(true);
+    backgroundMusic.play();
+    backgroundMusic.setVolume(
+      ClientSettings.musicVolume() / 100f * ClientSettings.masterVolume() / 100f);
+
+    ClientSettings.setOnVolumeChange(
+      (key, value) -> {
+        if (key.equals(ClientSettings.KEY_MUSIC_VOLUME)
+          || key.equals(ClientSettings.KEY_MASTER_VOLUME)) {
+          backgroundMusic.setVolume(
+            ClientSettings.musicVolume() / 100f * ClientSettings.masterVolume() / 100f);
+        }
+      });
   }
 }
